refactor: deduplicate run_git_command and admin-level area dicts (SU#563, group 2)
- Extract run_git_command to git/_utils.py; replace 4 identical copies
  in git_status, git_operations, branch_analyzer, git_workflow
- Extract ADMIN_LEVEL_AVG_AREA_KM2 and aliases to geo/_admin_areas.py;
  replace identical copies in h3_utils and s2_utils
- UserProfile collision in config/__init__.py verified already resolved
  via aliasing (EnhancedUserProfile, PydanticUserProfile)

Self-Review: .review/su563-refactors.md
Self-Review-Source: https://github.com/siege-analytics/siege_utilities/issues/563
