+++ siege_utilities/git/branch_analyzer.py
@@ -3,7 +3,6 @@ Branch analysis utilities for git repositories.
 Based on the Change-Agent-AI branch status generator.
 """
 
-import subprocess
 import re
 from datetime import datetime
 from pathlib import Path
@@ -11,24 +10,7 @@ from typing import List, Dict, Optional
 
 from siege_utilities.core.logging import log_info
 from siege_utilities.exceptions import GitError
-
-
-def run_git_command(*args, repo_path: str = ".", check: bool = True) -> str:
-    """Run a git command and return the output."""
-    try:
-        result = subprocess.run(
-            ["git"] + list(args),
-            cwd=repo_path,
-            capture_output=True,
-            text=True,
-            check=check,
-            timeout=30,  # writing-code:15: bounded git wait
-        )
-        return result.stdout.strip()
-    except subprocess.CalledProcessError as e:
-        if check:
-            raise GitError(f"Git command failed: {' '.join(args)} - {e.stderr}")
-        return ""
+from ._utils import run_git_command
 
 def analyze_branch_status(repo_path: str = ".") -> Dict[str, str]:
     """Analyze the current branch status."""
